PAPI upgrade + better link testcase
In new PAPI version, there's an API to decode call data from the
extrinsic.
However, it didn't work with the previous setup where we had multiple
networks with the same metadata.
But after thinking about it for a bit, it actually doesn't really make
sense to have duplicating papi declarations only for different WS urls,
so I've rewrote that part, which should be more efficient anyway.

A follow-up to #170